# improve CI caching of compiled dependencies