--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -56,8 +56,8 @@ jobs:
       - uses: actions/cache@v2
         if: matrix.sys == 'disable'
         with:
-          path: ports/archives
-          key: tarballs-ubuntu-${{hashFiles('**/dependencies.yml')}}
+          path: ports
+          key: ports-ubuntu-${{hashFiles('**/dependencies.yml')}}
       - run: bundle install --local || bundle install
       - run: bundle exec rake compile -- --${{matrix.sys}}-system-libraries
       - run: bundle exec rake test
@@ -77,8 +77,8 @@ jobs:
       - uses: actions/cache@v2
         if: matrix.sys == 'disable'
         with:
-          path: ports/archives
-          key: tarballs-ubuntu-${{hashFiles('**/dependencies.yml')}}
+          path: ports
+          key: ports-ubuntu-${{hashFiles('**/dependencies.yml')}}
       - run: bundle install --local || bundle install
       - run: bundle exec rake compile -- --${{matrix.sys}}-system-libraries
       - run: bundle exec rake test:valgrind
@@ -132,8 +132,8 @@ jobs:
       - uses: actions/cache@v2
         if: matrix.sys == 'disable'
         with:
-          path: ports/archives
-          key: tarballs-ubuntu-${{hashFiles('**/dependencies.yml')}}
+          path: ports
+          key: ports-ubuntu-${{hashFiles('**/dependencies.yml')}}
       - run: bundle install --local || bundle install
       - run: bundle exec rake compile -- --${{matrix.sys}}-system-libraries
       - run: bundle exec rake test
@@ -155,8 +155,8 @@ jobs:
       - uses: actions/cache@v2
         if: matrix.sys == 'disable'
         with:
-          path: ports/archives
-          key: tarballs-ubuntu-${{hashFiles('**/dependencies.yml')}}
+          path: ports
+          key: ports-ubuntu-${{hashFiles('**/dependencies.yml')}}
       - run: bundle install --local || bundle install
       - run: bundle exec rake compile -- --${{matrix.sys}}-system-libraries
       - run: bundle exec rake test:valgrind
@@ -178,8 +178,8 @@ jobs:
       - uses: actions/cache@v2
         if: matrix.sys == 'disable'
         with:
-          path: ports/archives
-          key: tarballs-macos-${{hashFiles('**/dependencies.yml')}}
+          path: ports
+          key: ports-macos-${{hashFiles('**/dependencies.yml')}}
       - run: bundle exec rake compile -- --${{matrix.sys}}-system-libraries
       - run: bundle exec rake test
 
@@ -205,8 +205,8 @@ jobs:
       - uses: actions/cache@v2
         if: matrix.sys == 'disable'
         with:
-          path: ports/archives
-          key: tarballs-windows-${{hashFiles('**/dependencies.yml')}}
+          path: ports
+          key: ports-windows-${{hashFiles('**/dependencies.yml')}}
       - run: bundle exec rake compile -- --${{matrix.sys}}-system-libraries
       - run: bundle exec rake test