**What problem is this PR intended to solve?**

Quite a bit of CI time is spent compiling our packaged dependencies (libxml2, libxslt, libiconv, zlib). For a given platform, we should be able to cache these libraries once compiled.

This PR does two things:

- move libgumbo out of `/ports`
- caches `/ports` for most jobs that were previously caching just `/ports/archives` (which only contains the tarballs)

I'll run CI twice on this branch and compare the timing to previous runs to make sure this is actually effective.