@flavorjones 2021-06-19
Pretty compelling before/after timings for the `rake compile` job:

|              | before | after |
|--------------|--------|-------|
| linux 2.7    | 1m12s  | 18s   |
| linux head   | 1m29s  | 17s   |
| macos head   | 2m10s  | 34s   |
| windows head | 9m47s  | 1m42s |